requirements.txt
jinja2
pandas
Pillow
python-slugify